+++ examples/base/src/pages/Dashboard/index.tsx
@@ -2,29 +2,34 @@ import React from 'react'
 import { Button } from 'uiw'
 import { useNavigate } from 'react-router-dom'
 import { useSWRConfig } from 'swr'
-import { request } from '@uiw-admin/utils'
+import { Dispatch, RootState } from '@uiw-admin/models'
+import { useSelector, useDispatch } from 'react-redux'
 
 const Dashboard = () => {
   const navigate = useNavigate()
-  const [state, setState] = React.useState('')
+  // const [state, setState] = React.useState('')
+
+  const dispatch = useDispatch<Dispatch>()
+
   const { cache } = useSWRConfig()
-
+  const stores = useSelector((state: RootState) => state)
+  console.log(stores)
   console.log('cache', cache.get('login'))
-
-  React.useEffect(() => {
-    request('/api', {
-      method: 'POST',
-      requestType: 'urlencoded',
-      body: {
-        a: 1,
-      },
-    })
-  }, [])
 
   return (
     <div>
       Dashboard
-      <input value={state} onChange={(event) => setState(event.target.value)} />
+      <input
+        value={stores.docDs.test}
+        onChange={(event) => {
+          dispatch({
+            type: 'docDs/updateState',
+            payload: {
+              test: event.target.value,
+            },
+          })
+        }}
+      />
       <hr />
       <Button onClick={() => navigate('/', { replace: true })}>Logout</Button>
     </div>