+++ examples/base/src/pages/Demo/Detail/index.tsx
@@ -4,7 +4,7 @@ import { Notify, Slider } from 'uiw'
 import { useSelector, useDispatch } from 'react-redux'
 import { RootState, Dispatch } from '@uiw-admin/models'
 import { insert, update } from '../../../servers/demo'
-import { items } from './items'
+import { items, items2 } from './items'
 import useSWR from 'swr'
 
 interface DetailProps {
@@ -16,13 +16,16 @@ const Detail = ({ updateData, onSearch }: DetailProps) => {
   const dispatch = useDispatch<Dispatch>()
   const [option, setOption] = React.useState<any>([])
   const [loading, setLoading] = React.useState(false)
+
   const {
-    demo: { drawerVisible, tableType, queryInfo, isView },
+    demo: { drawerVisible, tableType, queryInfo = {}, isView },
   } = useSelector((state: RootState) => state)
 
   const form = useForm()
+  const form2 = useForm()
   const onClose = () => dispatch({ type: 'demo/clean' })
 
+  // eslint-disable-next-line no-unused-vars
   const { mutate } = useSWR(
     [
       (tableType === 'add' && insert) || (tableType === 'edit' && update),
@@ -59,6 +62,18 @@ const Detail = ({ updateData, onSearch }: DetailProps) => {
       }, 2000)
     }
   }
+  const handleSave = async () => {
+    // 触发验证
+    await form?.submitvalidate()
+    await form2?.submitvalidate()
+    // 获取错误信息
+    const errors = form.getErrors()
+    const errors2 = form2.getErrors()
+
+    if (errors && Object.keys(errors).length > 0) return
+    if (errors2 && Object.keys(errors2).length > 0) return
+    mutate()
+  }
 
   return (
     <ProDrawer
@@ -74,7 +89,7 @@ const Detail = ({ updateData, onSearch }: DetailProps) => {
           type: 'danger',
           style: { width: 80 },
           show: !isView,
-          onClick: () => form.submitvalidate(),
+          onClick: handleSave,
         },
         {
           label: '取消',
@@ -94,22 +109,6 @@ const Detail = ({ updateData, onSearch }: DetailProps) => {
         }}
         formType={isView ? 'pure' : 'card'}
         readOnly={isView}
-        onSubmit={(
-          initial: Record<string, any>,
-          current: Record<string, any>
-        ) => {
-          const errorObj: any = {}
-          if (!current?.input) {
-            errorObj.input = '名字不能为空'
-          }
-          if (Object.keys(errorObj).length > 0) {
-            const err: any = new Error()
-            err.filed = errorObj
-            Notify.error({ title: '提交失败！' })
-            throw err
-          }
-          mutate()
-        }}
         // 更新表单的值
         onChange={(
           initial: Record<string, any>,
@@ -126,6 +125,28 @@ const Detail = ({ updateData, onSearch }: DetailProps) => {
             },
           }) as any
         }
+      />
+      <div style={{ marginTop: 10 }} />
+      <ProForm
+        form={form2}
+        title="个人信息"
+        customWidgetsList={{
+          slider: Slider,
+        }}
+        cardProps={{
+          noHover: true,
+        }}
+        formType={isView ? 'pure' : 'card'}
+        readOnly={isView}
+        // 更新表单的值
+        onChange={(
+          initial: Record<string, any>,
+          current: Record<string, any>
+        ) => {
+          updateData({ queryInfo: { ...queryInfo, ...current } })
+        }}
+        buttonsContainer={{ justifyContent: 'flex-start' }}
+        formDatas={items2(queryInfo) as any}
       />
     </ProDrawer>
   )