+++ examples/base/src/pages/Demo/Detail/items.tsx
@@ -2,20 +2,30 @@ import React from 'react'
 import { formatter } from 'uiw'
 export const items = (queryInfo: any, attr: any) => [
   {
-    label: '输入框',
+    label: '司机手机号',
     key: 'input',
     widget: 'input',
     initialValue: queryInfo?.input,
     widgetProps: {},
-    required: true,
-    // 提示信息
-    help: '输入框不能为空',
+    rulers: [
+      { required: true, message: '请输入' },
+      { pattern: new RegExp(/[1][3][0-9]{9}$/), message: '请输入正确手机号' },
+    ],
   },
   {
     label: '多行文本输入框',
     key: 'textarea',
     widget: 'textarea',
     initialValue: queryInfo?.textarea,
+    rulers: [
+      {
+        validator: (value: any | any[] = '') => {
+          if (!value) return false
+          return true
+        },
+        message: '请输入值',
+      },
+    ],
   },
   {
     label: '选择器',
@@ -28,12 +38,14 @@ export const items = (queryInfo: any, attr: any) => [
       { value: 4, label: '东北大冻梨' },
     ],
     initialValue: queryInfo?.select,
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '开关',
     key: 'switch',
     widget: 'switch',
     initialValue: queryInfo?.switch,
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '单选框',
@@ -44,6 +56,7 @@ export const items = (queryInfo: any, attr: any) => [
       { label: '女', value: 'girl' },
     ],
     initialValue: queryInfo?.radio,
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '多选框',
@@ -54,6 +67,7 @@ export const items = (queryInfo: any, attr: any) => [
       { label: '湖北菜', value: 'hubei' },
     ],
     initialValue: queryInfo?.checkbox,
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '年月日时分秒',
@@ -65,6 +79,7 @@ export const items = (queryInfo: any, attr: any) => [
     initialValue:
       queryInfo.dateInputsecond &&
       formatter('YYYY-MM-DD HH:mm:ss', new Date(queryInfo.dateInputsecond)),
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '年月日',
@@ -73,6 +88,7 @@ export const items = (queryInfo: any, attr: any) => [
     initialValue:
       queryInfo.dateInput &&
       formatter('YYYY-MM-DD', new Date(queryInfo.dateInput)),
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '年月',
@@ -81,12 +97,14 @@ export const items = (queryInfo: any, attr: any) => [
     initialValue:
       queryInfo.monthPicker &&
       formatter('YYYY-MM', new Date(queryInfo.monthPicker)),
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '时分秒',
     key: 'timePicker',
     widget: 'timePicker',
     initialValue: queryInfo.timePicker && new Date(queryInfo.timePicker),
+    rulers: [{ required: true, message: '请选择' }],
   },
   // 只读模式下支持读取React.ReactNode
   {
@@ -99,6 +117,7 @@ export const items = (queryInfo: any, attr: any) => [
     ) : (
       queryInfo?.slider
     ),
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: 'selectMultiple',
@@ -115,6 +134,7 @@ export const items = (queryInfo: any, attr: any) => [
       showSearch: true,
     },
     initialValue: queryInfo?.selectMultiple,
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '评分',
@@ -122,6 +142,7 @@ export const items = (queryInfo: any, attr: any) => [
     widget: 'rate',
     readSpan: 2,
     initialValue: queryInfo?.rate,
+    rulers: [{ required: true, message: '请选择' }],
   },
   {
     label: '上传组件',
@@ -133,8 +154,25 @@ export const items = (queryInfo: any, attr: any) => [
       uploadType: 'card',
       multiple: true,
       maxNumber: 2,
-      readonly: attr.isView,
+      showFileIcon: {
+        showPreviewIcon: true,
+        showRemoveIcon: !attr.isView,
+      },
     },
     initialValue: queryInfo?.upload,
+    rulers: [{ required: true, message: '请上传' }],
+  },
+]
+
+export const items2 = (queryInfo: any) => [
+  {
+    label: '手机号',
+    key: 'input2',
+    widget: 'input',
+    initialValue: queryInfo?.input2,
+    widgetProps: {},
+    rulers: [
+      { pattern: new RegExp(/[1][3][0-9]{9}$/), message: '请输入正确手机号' },
+    ],
   },
 ]