+++ packages/components/src/ProForm/formdom.tsx
@@ -1,28 +1,66 @@
-import React from 'react';
+import React, { useRef, useEffect } from 'react';
 import { Form, Button, Col, Row, FormFieldsProps } from 'uiw';
 import { ProFormProps } from './type';
 import { useStore } from './hooks/store';
+import { fromValidate } from './utils';
 import './style/form-item.less';
 
 function FormDom({
+  formDatas,
   formfields,
   onSubmit,
   onChange,
   buttonsContainer,
   showSaveButton = false,
   showResetButton = false,
-  saveButtonProps,
-  resetButtonProps,
+  saveButtonProps = {},
+  resetButtonProps = {},
 }: ProFormProps & {
   formfields: Record<string, FormFieldsProps<{}>> | undefined;
 }) {
+  const baseRef = useRef<any>();
   const store = useStore();
-  const { clickRef } = store as any;
+
+  const { formRef } = store as any;
+
+  useEffect(() => {
+    if (formRef && baseRef) {
+      formRef.current = {
+        ...formRef.current,
+        submitvalidate: baseRef.current?.onSubmit,
+        resetForm: baseRef.current?.resetForm,
+        getFieldValues: baseRef.current?.getFieldValues,
+      };
+    }
+  }, [baseRef.current]);
+
   return (
     <Form
+      ref={baseRef}
       style={{ background: '#fff', paddingBottom: 10, marginBottom: 14 }}
       resetOnSubmit={false}
-      onSubmit={({ initial, current }) => onSubmit?.(initial, current)}
+      onSubmit={({ initial, current }) => {
+        // 如果传入了onSubmit走onSubmit,否则主动验证
+        if (onSubmit) {
+          onSubmit?.(initial, current);
+        } else {
+          const validateList =
+            (formDatas &&
+              formDatas.length > 0 &&
+              formDatas.map((item) => ({
+                key: item.key,
+                value: current[item.key],
+                rulers: item.rulers,
+              }))) ||
+            [];
+          const errorObj = fromValidate(validateList);
+          if (Object.keys(errorObj).length > 0) {
+            const err: any = new Error();
+            err.filed = errorObj;
+            throw err;
+          }
+        }
+      }}
       onChange={({ initial, current }) => onChange?.(initial, current)}
       onSubmitError={(error) => {
         if (error.filed) {
@@ -33,6 +71,13 @@ function FormDom({
       fields={formfields}
     >
       {({ fields, state, canSubmit, resetForm }) => {
+        const { errors } = state;
+        if (formRef) {
+          formRef.current = {
+            ...formRef.current,
+            errors: errors,
+          };
+        }
         return (
           <React.Fragment>
             <Row gutter={10}>
@@ -46,32 +91,20 @@ function FormDom({
               })}
             </Row>
             <div className="w-form-item-center" style={{ ...buttonsContainer }}>
-              {showSaveButton ? (
-                <Button
-                  {...saveButtonProps}
-                  ref={clickRef}
-                  disabled={!canSubmit()}
-                  htmlType="submit"
-                >
-                  {saveButtonProps?.label || '提交'}
-                </Button>
-              ) : (
-                <Button
-                  style={{ display: 'none' }}
-                  ref={clickRef}
-                  disabled={!canSubmit()}
-                  htmlType="submit"
-                  {...saveButtonProps}
-                >
-                  {saveButtonProps?.label || '提交'}
-                </Button>
-              )}
+              <Button
+                style={{ display: showSaveButton ? 'flex' : 'none' }}
+                disabled={!canSubmit()}
+                htmlType="submit"
+                {...saveButtonProps}
+              >
+                {saveButtonProps.label || '提交'}
+              </Button>
               <Button
                 style={{ display: showResetButton ? 'flex' : 'none' }}
                 onClick={resetForm}
                 {...resetButtonProps}
               >
-                {resetButtonProps?.label || '重置'}
+                {resetButtonProps.label || '重置'}
               </Button>
             </div>
           </React.Fragment>