+++ CHANGELOG.md
@@ -4,11 +4,32 @@
 
 严格遵循 [`Semantic Versioning 2.0.0`](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
-## 5.2.13(未发版)
+
+## 5.2.14
+
+- @uiw-admin/components
+  - 🐞`ProForm`: 增加rulers验证
+  - 🐞`ProForm`: `useFormProps`删除clickRef & 新增getErrors获取错误信息 & getFormValues更名getFieldValues(与uiw保持一致)
+  - 完善upload组件
+
+- @uiw-admin/basic-layouts
+  - 🌟 退出清空token
+- @uiw-admin/config 
+  - 🌟 新增 `rematch` 属性 
+- @uiw-admin/plugins 
+  - 🌟 动态生成路由映射`model`文件
+  - 🐞 `rematch` 监听文件变化问题
+
+## 5.2.13
+
 - @uiw-admin/components
   - 🐞`ProForm`: 新增Rate组件
+  - 🐞`ProForm`: `form` api增加resetForm 和 formRef 和 getFormValues返回值
+- @uiw-admin/utils
+  - 🌟 `request`修复`body`参数问题
 
 ## 5.2.12
+
 - @uiw-admin/components
   - 🐞`ProTable`: 修复分页错误